break glob extensions into its own module
